+++ sei-ibc-go/testing/mock/mock.go
@@ -3,6 +3,7 @@ package mock
 import (
 	"bytes"
 	"encoding/json"
+	"strconv"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -27,12 +28,14 @@ const (
 )
 
 var (
-	MockAcknowledgement      = channeltypes.NewResultAcknowledgement([]byte("mock acknowledgement"))
-	MockFailAcknowledgement  = channeltypes.NewErrorAcknowledgement("mock failed acknowledgement")
-	MockPacketData           = []byte("mock packet data")
-	MockFailPacketData       = []byte("mock failed packet data")
-	MockAsyncPacketData      = []byte("mock async packet data")
-	MockCanaryCapabilityName = "mock canary capability name"
+	MockAcknowledgement             = channeltypes.NewResultAcknowledgement([]byte("mock acknowledgement"))
+	MockFailAcknowledgement         = channeltypes.NewErrorAcknowledgement("mock failed acknowledgement")
+	MockPacketData                  = []byte("mock packet data")
+	MockFailPacketData              = []byte("mock failed packet data")
+	MockAsyncPacketData             = []byte("mock async packet data")
+	MockRecvCanaryCapabilityName    = "mock receive canary capability name"
+	MockAckCanaryCapabilityName     = "mock acknowledgement canary capability name"
+	MockTimeoutCanaryCapabilityName = "mock timeout canary capability name"
 )
 
 var _ porttypes.IBCModule = AppModule{}
@@ -194,7 +197,12 @@ func (am AppModule) OnChanCloseConfirm(sdk.Context, string, string) error {
 // OnRecvPacket implements the IBCModule interface.
 func (am AppModule) OnRecvPacket(ctx sdk.Context, packet channeltypes.Packet, relayer sdk.AccAddress) exported.Acknowledgement {
 	// set state by claiming capability to check if revert happens return
-	am.scopedKeeper.NewCapability(ctx, MockCanaryCapabilityName)
+	_, err := am.scopedKeeper.NewCapability(ctx, MockRecvCanaryCapabilityName+strconv.Itoa(int(packet.GetSequence())))
+	if err != nil {
+		// application callback called twice on same packet sequence
+		// must never occur
+		panic(err)
+	}
 	if bytes.Equal(MockPacketData, packet.GetData()) {
 		return MockAcknowledgement
 	} else if bytes.Equal(MockAsyncPacketData, packet.GetData()) {
@@ -205,11 +213,25 @@ func (am AppModule) OnRecvPacket(ctx sdk.Context, packet channeltypes.Packet, re
 }
 
 // OnAcknowledgementPacket implements the IBCModule interface.
-func (am AppModule) OnAcknowledgementPacket(sdk.Context, channeltypes.Packet, []byte, sdk.AccAddress) error {
+func (am AppModule) OnAcknowledgementPacket(ctx sdk.Context, packet channeltypes.Packet, _ []byte, _ sdk.AccAddress) error {
+	_, err := am.scopedKeeper.NewCapability(ctx, MockAckCanaryCapabilityName+strconv.Itoa(int(packet.GetSequence())))
+	if err != nil {
+		// application callback called twice on same packet sequence
+		// must never occur
+		panic(err)
+	}
+
 	return nil
 }
 
 // OnTimeoutPacket implements the IBCModule interface.
-func (am AppModule) OnTimeoutPacket(sdk.Context, channeltypes.Packet, sdk.AccAddress) error {
+func (am AppModule) OnTimeoutPacket(ctx sdk.Context, packet channeltypes.Packet, _ sdk.AccAddress) error {
+	_, err := am.scopedKeeper.NewCapability(ctx, MockTimeoutCanaryCapabilityName+strconv.Itoa(int(packet.GetSequence())))
+	if err != nil {
+		// application callback called twice on same packet sequence
+		// must never occur
+		panic(err)
+	}
+
 	return nil
 }